requirements.txt
# === Core Python Libraries ===
python-dotenv==1.0.1
requests==2.31.0
ccxt==4.3.59

# === Optional Tools (used by Midas dashboard / plotting) ===
matplotlib==3.8.0
pandas==2.2.3

# === Notes ===
# - No Google Sheets dependencies (fully removed)
# - Works on Render, PyCharm, and local Windows/Linux/Mac
# - Tested with Python 3.10 – 3.13
